# Conflicts:
#	src/bddTest/java/pageObjects/SM/searchPage/SearchPageImpl.java
#	src/bddTest/resources/features/SM/SuperMatchSearch.feature

+++ src/bddTest/java/pageObjects/SP/adminPages/HomePageImpl.java
@@ -8,6 +8,8 @@ import org.openqa.selenium.support.ui.ExpectedConditions;
 import pageObjects.COMMON.GlobalSearch;
 import pageObjects.COMMON.PageObjectFacadeImpl;
 
+import java.util.List;
+
 public class HomePageImpl extends PageObjectFacadeImpl {
 
     private Logger logger;
@@ -15,6 +17,12 @@ public class HomePageImpl extends PageObjectFacadeImpl {
 
     public HomePageImpl() {
         logger = Logger.getLogger(HomePageImpl.class);
+    }
+
+    public void verifyUserIsLoggedInForNoRole() {
+        //Check the signout button is present for No Role
+        Assert.assertTrue("User did not signed in",signoutButtonForNoRole().isDisplayed());
+        logger.info("Logged in successfully");
     }
 
     public void verifyUserIsLoggedIn() {
@@ -29,11 +37,21 @@ public class HomePageImpl extends PageObjectFacadeImpl {
     }
 
     public void logout() {
-        link(By.id("user-dropdown")).click();
-        driver.findElement(By.cssSelector("div[id='user-dropdown-signout']")).click();
-        waitUntilPageFinishLoading();
-        Assert.assertTrue(getDriver().getCurrentUrl().contains("login"));
-        driver.manage().deleteAllCookies();
+        try{
+            List<WebElement> signoutButton = driver.findElements(By.cssSelector("button[class='ui mini button']"));
+            if (signoutButton.size()==1){
+                signoutButtonForNoRole().click();
+                waitUntilPageFinishLoading();
+                Assert.assertTrue(getDriver().getCurrentUrl().contains("login"));
+                driver.manage().deleteAllCookies();
+            }
+        }catch(Exception e){
+            link(By.id("user-dropdown")).click();
+            driver.findElement(By.cssSelector("div[id='user-dropdown-signout']")).click();
+            waitUntilPageFinishLoading();
+            Assert.assertTrue(getDriver().getCurrentUrl().contains("login"));
+            driver.manage().deleteAllCookies();
+        }
     }
 
     public String selectTheFistInstitutionOnTheList() {
@@ -89,10 +107,12 @@ public class HomePageImpl extends PageObjectFacadeImpl {
         waitUntilPageFinishLoading();
     }
 
+    private WebElement userDropdownSingout() { return button(By.id("user-dropdown-signout"));}
     private WebElement userDropdown() {
         return button(By.id("user-dropdown"));
     }
     private WebElement userListTable() {
         return button(By.cssSelector("[class='ui table']"));
     }
+    private WebElement signoutButtonForNoRole(){return driver.findElement(By.cssSelector("button[class='ui mini button']")); }
 }
